+++ src/main/java/com/mahait/gov/in/repository/DdoBillGroupRepo.java
@@ -43,7 +43,7 @@ public interface DdoBillGroupRepo {
 
 	public int deleteEmpMpgDdoAllowDeduc(String sevaarthId);
 
-	public int saveEmpMpgDdoAllowDeduc(Object allow_deduct_id, int department_id, int empId, String sevaarthId,
+	public int saveEmpMpgDdoAllowDeduc(Object allow_deduct_id, Long department_id, Long empId, String sevaarthId,
 			String effectiveDate);
 
 	public List<Object[]> empEligibilityForAllowAndDeductCheckBoxId(String id);
